+++ src/StreetMap.h
@@ -26,9 +26,9 @@ using namespace std;
 #define EARTH_RADIUS 6371.0
 
 class StreetMap {
-	map<unsigned long int,Node> nodes;
-	map<unsigned long int,Road> roads;
-	Graph<unsigned long int> graph;
+	map<int,Node> nodes;
+	map<int,Road> roads;
+	Graph<int> graph;
 	//vector<Road> roads;
 
 public:
